feat: corrige movimentacao em montagem de fita
