+++ src/robo/CLI/clear_states.py
@@ -1,0 +1,4 @@
+def clear_states(robo):
+    robo.get_alarm_state()
+    robo.clear_all_alarms()
+    robo.home()
