+++ src/robo/CLI/controle_manual.py
@@ -1,0 +1,41 @@
+import keyboard
+
+def controle_manual(robo, delta=20):
+    """Controle manual com movimentos do tipo MOVJ"""
+    print("Modo de controle manual ativado (MOVJ). Use as teclas:")
+    print("  ←/→: mover em X | ↑/↓: mover em Y | W/S: mover em Z")
+    print("  A/D: ajustar rotação | Q: sair")
+    
+    while True:
+        if keyboard.is_pressed('q'):
+            print("Saindo do modo de controle manual.")
+            break
+
+        if keyboard.is_pressed('h'):
+            robo.home();
+        
+
+        x, y, z, r, *_ = robo.pose()
+        dx = dy = dz = dr = 0
+        
+        # A fim de melhorar um pouco a fluidez, o que é feito
+        # é basicamente aumentar a variação de movimento
+        # a cada momento em que a tecla está pressionada.
+        # No final, passamos a posição atualizada ao robô
+        if keyboard.is_pressed('left'): dx = -delta
+        if keyboard.is_pressed('right'): dx = delta
+        if keyboard.is_pressed('up'): dy = delta
+        if keyboard.is_pressed('down'): dy = -delta
+        if keyboard.is_pressed('w'): dz = delta
+        if keyboard.is_pressed('s'): dz = -delta
+        if keyboard.is_pressed('a'): dr = -delta
+        if keyboard.is_pressed('d'): dr = delta
+
+        if dx or dy or dz or dr:
+            robo.movej_to(
+                x + dx,
+                y + dy,
+                z + dz,
+                r + dr,
+                wait=False
+            )
