+++ src/robo/pontos/pontos.json
@@ -108,7 +108,7 @@
           "suctionCup": "off"
         },
         {
-          "ponto": 1,
+          "ponto": 2,
           "x": 258.3503,
           "y": -102.9024,
           "z": 65.2056,
@@ -117,16 +117,16 @@
           "suctionCup": "on"
         },
         {
-          "ponto": 2,
+          "ponto": 3,
           "x": 255.452,
           "y": -105.5703,
           "z": 16.8745,
           "r": -22.4538,
-          "movimento": "movj",  
+          "movimento": "movl",  
           "suctionCup": "on"
         },
         {
-          "ponto": 3,
+          "ponto": 4,
           "x": 249.2568,
           "y": -99.2805,
           "z": 156.4643,
@@ -135,7 +135,7 @@
           "suctionCup": "on"
         },
         {
-          "ponto": 4,
+          "ponto": 5,
           "x": 18.4174,
           "y": -304.799,
           "z": 156.4643,
@@ -144,7 +144,7 @@
           "suctionCup": "on"
         },
         {
-          "ponto": 5,
+          "ponto": 6,
           "x": 11.0435,
           "y": -310.6665,
           "z": 144.9234,
@@ -153,7 +153,7 @@
           "suctionCup": "on"
         },
         {
-          "ponto": 6,
+          "ponto": 7,
           "x": 15.0558,
           "y": -311.9232,
           "z": -22.601,
@@ -162,7 +162,7 @@
           "suctionCup": "off"
         },
         {
-          "ponto": 7,
+          "ponto": 8,
           "x": 10.2297,
           "y": -290.9894,
           "z": 146.983,
